better date parsing for publications - delete unmaintainable tooltip code - saved interaction tooltip code for later - interaction tooltips won't work temporarily
